progress

+++ src/Order.cpp
@@ -1,5 +1,5 @@
 #include "Order.hpp"
 
-void Order::update_quantity(double act_volume){
+void Order::update_quantity(long act_volume){
     volume = act_volume;
 }
